+++ src/main/java/com/lhh/community/dto/QuestionDTO.java
@@ -12,7 +12,7 @@ import lombok.Data;
 @Data
 public class QuestionDTO {
 
-    private Integer id;
+    private Long id;
 
     private String title;
 
@@ -20,13 +20,13 @@ public class QuestionDTO {
 
     private Long gmtModified;
 
-    private Integer creator;
+    private Long creator;
 
-    private Integer commentCount;
+    private Long commentCount;
 
-    private Integer viewCount;
+    private Long viewCount;
 
-    private Integer likeCount;
+    private Long likeCount;
 
     private String tag;
