+++ src/main/java/com/lhh/community/entity/Question.java
@@ -7,7 +7,7 @@ import java.io.Serializable;
 @Data
 public class Question implements Serializable {
 
-    private Integer id;
+    private Long id;
 
     private String title;
 
@@ -15,13 +15,13 @@ public class Question implements Serializable {
 
     private Long gmtModified;
 
-    private Integer creator;
+    private Long creator;
 
-    private Integer commentCount;
+    private Long commentCount;
 
-    private Integer viewCount;
+    private Long viewCount;
 
-    private Integer likeCount;
+    private Long likeCount;
 
     private String tag;
