+++ src/main/java/com/lhh/community/entity/User.java
@@ -6,7 +6,8 @@ import java.io.Serializable;
 
 @Data
 public class User implements Serializable {
-    private Integer id;
+
+    private Long id;
 
     private String name;
